Moved PDS date converter to registry-commons project
