+++ src/main/java/gov/nasa/pds/harvest/meta/AutogenExtractor.java
@@ -16,8 +16,8 @@ import org.w3c.dom.NodeList;
 import gov.nasa.pds.harvest.Constants;
 import gov.nasa.pds.harvest.cfg.model.AutogenCfg;
 import gov.nasa.pds.harvest.util.FieldMap;
-import gov.nasa.pds.harvest.util.date.PdsDateConverter;
 import gov.nasa.pds.harvest.util.xml.XmlDomUtils;
+import gov.nasa.pds.registry.common.util.date.PdsDateConverter;
 
 
 public class AutogenExtractor
