+++ src/main/java/gov/nasa/pds/harvest/meta/XPathExtractor.java
@@ -3,15 +3,22 @@ package gov.nasa.pds.harvest.meta;
 import org.w3c.dom.Document;
 
 import gov.nasa.pds.harvest.util.FieldMap;
-import gov.nasa.pds.harvest.util.date.PdsDateConverter;
 import gov.nasa.pds.harvest.util.xml.XPathCache;
 import gov.nasa.pds.harvest.util.xml.XPathUtils;
+import gov.nasa.pds.registry.common.util.date.PdsDateConverter;
 
 
+/**
+ * Extract data from PDS4 label by XPath
+ * @author karpenko
+ */
 public class XPathExtractor
 {
     private PdsDateConverter dateConverter;
     
+    /**
+     * Constructor
+     */
     public XPathExtractor()
     {
         dateConverter = new PdsDateConverter(true);
